+++ ssr/extable-go/escape.go
@@ -1,0 +1,14 @@
+package extable
+
+import "strings"
+
+func escapeHTML(text string) string {
+	replacer := strings.NewReplacer(
+		"&", "&amp;",
+		"<", "&lt;",
+		">", "&gt;",
+		"\"", "&quot;",
+		"'", "&#39;",
+	)
+	return replacer.Replace(text)
+}